Moved all the images and svg to assets
